+++ node/src/types/kernel/api/notebook/removeNotebook/response.d.ts
@@ -15,9 +15,17 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-
-/* directories */
-
-/* flles */
-export * from "./payload";
-export * from "./response";
+/**
+ * Remove a notebook
+ */
+export interface IResponse {
+    /**
+     * status code
+     */
+    code: number;
+    data: null;
+    /**
+     * status message
+     */
+    msg: string;
+}