+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - 添加 API `/api/notebook/openNotebook` | Add API `/api/notebook/openNotebook`
 - 添加 API `/api/notebook/renameNotebook` | Add API `/api/notebook/renameNotebook`
 - 添加 API `/api/notebook/createNotebook` | Add API `/api/notebook/createNotebook`
+- 添加 API `/api/notebook/removeNotebook` | Add API `/api/notebook/removeNotebook`
 
 ## v0.0.4 / 2023-06-29